done with crud in blogs

+++ src/handlers/handlers.ts
@@ -3,7 +3,7 @@ import { BlogType, CommentType, UserType } from "../schema/schema";
 import User from "../models/User";
 import Blog from "../models/Blog";
 import Comment from "../models/Comment";
-import { Document } from 'mongoose';
+import { Document, startSession } from 'mongoose';
 import { compareSync, hashSync } from 'bcryptjs';
 
 type DocumentType = Document<any, any, any>
@@ -96,14 +96,25 @@ const mutations = new GraphQLObjectType({
                 title: { type: new GraphQLNonNull(GraphQLString) },
                 content: { type: new GraphQLNonNull(GraphQLString) },
                 date: { type: new GraphQLNonNull(GraphQLString) },
+                user: { type: new GraphQLNonNull(GraphQLID) },
             },
-            async resolve(parent, { title, content, date }) {
+            async resolve(parent, { title, content, date, user }) {
                 let blog: DocumentType
+                const session = await startSession()
                 try {
-                    blog = new Blog({ title, content, date });
-                    return await blog.save();
+                    session.startTransaction({ session });
+                    blog = new Blog({ title, content, date, user });
+                    const existingUser = await User.findById(user)
+                    if (!existingUser) {
+                        return new Error("User Not Found")
+                    }
+                    existingUser.blogs.push(blog);
+                    await existingUser.save({ session })
+                    return await blog.save({ session });
                 } catch (error) {
                     return new Error(error);
+                } finally {
+                    await session.commitTransaction()
                 }
             }
         },
@@ -143,11 +154,20 @@ const mutations = new GraphQLObjectType({
             },
             async resolve(parent, { id }) {
                 let existingBlog: DocumentType
+                const session = await startSession()
                 try {
-                    existingBlog = await Blog.findById(id)
+                    session.startTransaction({ session });
+                    existingBlog = await Blog.findById(id).populate("user");
+                    //@ts-ignore
+                    const existingUser = existingBlog?.user;
+                    if (!existingUser) {
+                        return new Error("No User linked with this blog")
+                    }
                     if (!existingBlog) {
                         return new Error("No blog Found")
                     }
+                    existingUser.blogs.pull(existingBlog);
+                    await existingUser.save({ session });
                     return await Blog.findByIdAndRemove(id)
                 } catch (error) {
                     return new Error(error)
